cluster introspection
