--- src/Data/Morpheus/Resolve/Generics/UnionRep.hs
@@ -1,30 +1,0 @@
-{-# LANGUAGE ExistentialQuantification #-}
-{-# LANGUAGE FlexibleContexts          #-}
-{-# LANGUAGE FlexibleInstances         #-}
-{-# LANGUAGE MultiParamTypeClasses     #-}
-{-# LANGUAGE PolyKinds                 #-}
-{-# LANGUAGE RankNTypes                #-}
-{-# LANGUAGE ScopedTypeVariables       #-}
-{-# LANGUAGE TypeApplications          #-}
-{-# LANGUAGE TypeOperators             #-}
-
-module Data.Morpheus.Resolve.Generics.UnionRep
-  ( UnionRep(..)
-  ) where
-
-import           Data.Morpheus.Types.Internal.Data       (DataField, DataTypeLib)
-import           Data.Morpheus.Types.Internal.Validation (SchemaValidation)
-import           Data.Proxy                              (Proxy (..))
-import           GHC.Generics
-
-class UnionRep f where
-  possibleTypes :: Proxy f -> [(DataField (), DataTypeLib -> SchemaValidation DataTypeLib)]
-
-instance UnionRep f => UnionRep (M1 D x f) where
-  possibleTypes _ = possibleTypes (Proxy @f)
-
-instance UnionRep f => UnionRep (M1 C x f) where
-  possibleTypes _ = possibleTypes (Proxy @f)
-
-instance (UnionRep a, UnionRep b) => UnionRep (a :+: b) where
-  possibleTypes _ = possibleTypes (Proxy @a) ++ possibleTypes (Proxy @b)
